backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
websockets==13.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
requests==2.32.3
pillow==11.0.0
bcrypt==4.0.1
email-validator==2.2.0

# PostgreSQL driver
psycopg2-binary==2.9.9

# Redis for real-time communication
redis==5.0.1

# System monitoring
psutil==6.1.0
aioredis==2.0.1

# AI Engine - YOLOv8 for Object Detection
torch==2.4.0
torchvision==0.19.0
ultralytics==8.0.196
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.3
